backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
python-multipart==0.0.9
rq==1.16.2
redis==5.0.4
boto3==1.34.131
slowapi==0.1.9
httpx==0.27.0
pytest==8.2.2
pytest-asyncio==0.23.7
anyio==4.4.0
scipy>=1.13.0
pandas>=2.2.0
numpy>=1.26.0
